fix: address review feedback on warning infrastructure
- Separate ParameterErrorCode::NONE from error cases in
  classify_error_code with comment explaining it is unreachable
- Add comment to ros2_medkit_msgs explaining why ROS2MedkitWarnings
  is not included (rosidl-generated code)
- Simplify format specifier in fault_manager_node: use %ld directly
  instead of PRId64 + static_cast (declare_parameter<int> returns long)

+++ src/ros2_medkit_fault_manager/src/fault_manager_node.cpp
@@ -17,7 +17,6 @@
 #include <yaml-cpp/yaml.h>
 
 #include <cctype>
-#include <cinttypes>
 #include <filesystem>
 #include <fstream>
 #include <nlohmann/json.hpp>
@@ -111,8 +110,7 @@ FaultManagerNode::FaultManagerNode(const rclcpp::NodeOptions & options) : Node("
   }
   auto max_snapshots = declare_parameter<int>("snapshots.max_per_fault", 10);
   if (max_snapshots < 0) {
-    RCLCPP_WARN(get_logger(), "snapshots.max_per_fault should be >= 0, got %" PRId64 ". Disabling limit.",
-                static_cast<int64_t>(max_snapshots));
+    RCLCPP_WARN(get_logger(), "snapshots.max_per_fault should be >= 0, got %ld. Disabling limit.", max_snapshots);
     max_snapshots = 0;
   }
